+++ apps/alu-learn/src/components/SaveLocation/index.ts
@@ -1,0 +1,2 @@
+import SaveLocation from "./SaveLocation";
+export default SaveLocation;