+++ ecosystem.config.js
@@ -6,8 +6,8 @@ module.exports = {
     args   : "start",
   }, {
     name   : "alu-cron",
-    script : "node /home/aluadmin/aludir3/apps/alu-learn/cron/_cron.mjs",
-    // script : "node /home/aluadmin/aludir2/cron/_cron.mjs",
+    // script : "node /home/aluadmin/aludir3/apps/alu-learn/cron/_cron.mjs",
+    script : "node /home/aluadmin/aludir2/cron/_cron.mjs",
     cwd    : "/home/aluadmin/aludir2",
   }],
 };